Move all files inside of charbee_byte module
